Solved conflicts
